requirements.txt
asgiref==3.5.0
certifi==2020.6.20
dj-database-url==0.5.0
Django==4.0.4
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2022.1
sqlparse==0.4.2
tk==0.1.0
whitenoise==6.0.0
gunicorn
django_cors_headers==3.11.0
